@@ -5,7 +5,7 @@ import '../widgets/auto_refresh.dart';
 import '../widgets/empty_card.dart';
 
 class CardColumnScreen extends StatefulWidget {
-  CardColumnScreen({Key? key}) : super(key: key);
+  const CardColumnScreen({super.key});
 
   @override
   _CardColumnScreenState createState() => _CardColumnScreenState();
@@ -17,6 +17,10 @@ class _CardColumnScreenState extends State<CardColumnScreen> {
     return AutoRefresh(
       duration: const Duration(milliseconds: 2000),
       child: Scaffold(
+        appBar: AppBar(
+          centerTitle: true,
+          title: const Text('Animated Column View'),
+        ),
         body: SafeArea(
           child: SingleChildScrollView(
             padding: const EdgeInsets.all(16.0),
@@ -37,26 +41,26 @@ class _CardColumnScreenState extends State<CardColumnScreen> {
                       padding: const EdgeInsets.symmetric(vertical: 8.0),
                       child: Row(
                         mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                        children: [
-                          const EmptyCard(height: 50.0, width: 50.0),
-                          const EmptyCard(height: 50.0, width: 50.0),
-                          const EmptyCard(height: 50.0, width: 50.0),
+                        children: const [
+                          EmptyCard(height: 50.0, width: 50.0),
+                          EmptyCard(height: 50.0, width: 50.0),
+                          EmptyCard(height: 50.0, width: 50.0),
                         ],
                       ),
                     ),
                     Row(
-                      children: [
-                        const Flexible(child: EmptyCard(height: 150.0)),
-                        const Flexible(child: EmptyCard(height: 150.0)),
+                      children: const [
+                        Flexible(child: EmptyCard(height: 150.0)),
+                        Flexible(child: EmptyCard(height: 150.0)),
                       ],
                     ),
                     Padding(
                       padding: const EdgeInsets.symmetric(vertical: 8.0),
                       child: Row(
-                        children: [
-                          const Flexible(child: EmptyCard(height: 50.0)),
-                          const Flexible(child: EmptyCard(height: 50.0)),
-                          const Flexible(child: EmptyCard(height: 50.0)),
+                        children: const [
+                          Flexible(child: EmptyCard(height: 50.0)),
+                          Flexible(child: EmptyCard(height: 50.0)),
+                          Flexible(child: EmptyCard(height: 50.0)),
                         ],
                       ),
                     ),
