@@ -1,3 +1,4 @@
+
 import 'package:flutter/material.dart';
 import 'package:flutter_staggered_animations/flutter_staggered_animations.dart';
 
@@ -5,7 +6,7 @@ import '../widgets/auto_refresh.dart';
 import '../widgets/empty_card.dart';
 
 class CardListScreen extends StatefulWidget {
-  CardListScreen({Key? key}) : super(key: key);
+  const CardListScreen({super.key});
 
   @override
   _CardListScreenState createState() => _CardListScreenState();
@@ -17,6 +18,10 @@ class _CardListScreenState extends State<CardListScreen> {
     return AutoRefresh(
       duration: const Duration(milliseconds: 2000),
       child: Scaffold(
+        appBar: AppBar(
+          centerTitle: true,
+          title: const Text('Animated List View'),
+        ),
         body: SafeArea(
           child: AnimationLimiter(
             child: ListView.builder(
@@ -27,7 +32,8 @@ class _CardListScreenState extends State<CardListScreen> {
                   position: index,
                   duration: const Duration(milliseconds: 375),
                   child: SlideAnimation(
-                    verticalOffset: 44.0,
+                    verticalOffset: index % 2 == 0 ? 88.0 : 88.0,
+                    horizontalOffset: index % 2 == 0 ? 88.0 : -88.0,
                     child: FadeInAnimation(
                       child: EmptyCard(
                         width: MediaQuery.of(context).size.width,
