@@ -1,3 +1,5 @@
+import 'dart:math';
+
 import 'package:flutter/material.dart';
 
 class EmptyCard extends StatelessWidget {
@@ -5,21 +7,27 @@ class EmptyCard extends StatelessWidget {
   final double? height;
 
   const EmptyCard({
-    Key? key,
+    super.key,
     this.width,
     this.height,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
+    final color = Random.secure();
+    final r = color.nextInt(255);
+    final g = color.nextInt(255);
+    final b = color.nextInt(255);
+
+    final randomCol = Color.fromARGB(225, r, g, b);
     return Container(
       width: width,
       height: height,
       margin: const EdgeInsets.symmetric(vertical: 8.0, horizontal: 8.0),
-      decoration: const BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.all(Radius.circular(4.0)),
-        boxShadow: <BoxShadow>[
+      decoration: BoxDecoration(
+        color: randomCol,
+        borderRadius: const BorderRadius.all(Radius.circular(4.0)),
+        boxShadow: const <BoxShadow>[
           BoxShadow(
             color: Colors.black12,
             blurRadius: 4.0,
