@@ -1,4 +1,5 @@
 import 'package:flutter/widgets.dart';
+
 import 'animation_configuration.dart';
 import 'animation_executor.dart';
 
@@ -8,11 +9,11 @@ class AnimationConfigurator extends StatelessWidget {
   final Widget Function(Animation<double>) animatedChildBuilder;
 
   const AnimationConfigurator({
-    Key? key,
+    super.key,
     this.duration,
     this.delay,
     required this.animatedChildBuilder,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
